unpinned_requirements.txt
python-docx
typer
PyPDF2
textract
pandas
